Use a smaller buffer when marshaling bonded<T>
When bonded<T> is serialized to untagged protocols like Simple Binary
it is first marshaled into a separate buffer using a tagged protocol and
then written as length prefixed blob to the untagged protocol.
OutputBuffer used for marshaling by default allocates 64KB which is
costly if the bonded<T> payload is small.