Fix bug in JSON parsing when string value is a date
When reading string fields, the SimpleJsonParser expects to find a JsonToken of
type String. But if the string value matches a date format, the actual JsonToken
that is seen is of type Date. To fix this, we are setting the JsonTextReader to not
try to parse strings as dates.

Also improve the exception text when reading unenxpected input to help know
what JsonToken was actually read, and fix a similar possible issue with parsing
doubles.