Add support for Deserializer without inlining
When the schemas dependency graph complexity increases, the expression
trees that are generated by the DeserializeTransform can become very
large, leading to long compilation times and high memory consumption.

This change adds some Deserializer constructors that support a flag to
disable expression inlining. As a result the runtime deserialization
operation will be slightly slower (with some added overhead of function
calls), but the initialization of the Deserializer is significantly
faster.