Add code generation tests
The tests compare checked-in generated files with code generated using
the current compiler.

Discrepancy indicates a change in the generated code which may be a
bug or an intentional change. In the latter case the generated file(s)
should be re-generated and updated in the same commit as the change.

Note that the checked-in files are generated with --no-banner option.