Don't initialize BondCodegen.Options in .targets
The BondCodegen.Options property (the per-file Bond codegen options)
should to be initialized. The build should either uses the global
options (the BondOptions property) to compile multiple schema files in a
batch, or the per-file Options property to compile individual files that
need different options. Initializing the BondCodegen.Options to
BondOptions breaks the conditional which determines which schema files
can be compiled as one batch.