Eliminate Marshaler/Serializer code duplication
Marshaler is a thin wrapper over Serializer adding 4 byte header
specifying protocol and its version. While the previous implementation
of Marshaler reused the Serializer source code, it resulted in a lot of
duplication of executable code if both Serializer and Marshaler were
instantiated for the same type(s). With this change Apply for Marshaler
simply delegates to Apply for Serializer.