Use fast pass-through for bonded<T, Reader&>
Fast pass-through avoids field-by-field transcoding when source and
destination protocol are the same. Instead it just chains the source
blob into the serialization output.

Previously fast pass-through was hooked only for transcoding using
bonded<T> but not on bonded<T, Reader&>. In addition to improving
performance this change decreases build-time and size of executable code
for the generated _apply.cpp which defines transcoding overloads for
Apply with bonded<T, Reader&>.