Don't cast the instance contained in Bonded<T>
Bonded<T> is an implementation of IBonded<T> which contains an instance
of type T. When it is converted or deserialized to another type U we
should not cast the object instance.

Deserialization is implemented using cloning which doesn't require any
cast at all because Clone<U>.From<T> doesn't require that U and T are
the same type.

For conversion to IBonded<U> instead of creating a new Bonded<T> from
casted object instance we can cast the Bonded<T> object to IBonded<U>.

This fixes bug #71.