Fix containers of nullable<T> and blob in Python
Custom converters from a C++ type to a Python type are not invoked when
accessing container elements by reference (the default access policy).
This lead to an error when accessing elements of a container of
nullable<T> or blob. The fix is to return elements by value for types
that have a converter.