Don't allow JSON null for scalar types
Bond value nothing does not have a wire representation. Optional fields
set to nothing are omitted during serialization and required fields
cause exception.

Simple JSON parser was incorrectly returning null as a value when a
scalar was expected. This appeared to work during Deserialization for
fields with default nothing because they are represented by Nullable in
C#. However that was only a lucky accident of the Deserializer
implementation. Transforms in general are not expected to be able to
handle null as result from IParser.Scalar.