Allow the word 'Schema' in namespace identifiers
The word 'Schema' is a reserved name in Bond lexer because it conflicts
with the Schema struct in the generated C++ code. In order to support
some legacy schema definitions we are relaxing the restriction, allowing
the word to be used as part of namespace identifiers. The word Schema in
a namespace identifier does not conflict with the Schema struct in the
generated C++ code.