Add cabal constraint: process < 1.4
Unfortunately the conflict between optparse-applicative, which sets an
upper bound on the version the process package dependency and cmdargs
which only has a lower bound is going to reappear on every release of
process.

The problem happens when gbc dependencies are first installed in the
sandbox and then we try to build gbc-tests using the same sandbox.