Fix logic determining which C++ headers to include
The C++ codegen failed to detect the need to generate #include <blob.h>
when the only use of blob was a field with default `nothing`.

For aliases of types that require extra header files the trigger should
be the alias usage rather than alias declaration.

Types of fields of base schema should not be considered.