Fix buffer corruption in InputStream
When reading a blob from InputStream the underlying buffer must be
marked as not-reusable so that subsequent reads from the stream don't
overwrite the blob content. The bug was that we were resetting the
canReuseBuffer flag *before* reading the blob. If reading resulted in
allocation of a new buffer the canReuseBuffer flag would be set to true.

Fixes #114.