Avoid conflicts with macro named U in nullable
Some consumers of Bond use other libraries that define a function-style
macro named U. This macro conflicts with nullable's use of U() when
constructing values. We now use the type name ValueType instead of U and
intentionally construct a local that we return to avoid triggering
function-style macro expansion.

I looked at the generated code for nullable<string>::make_value() that a
release build with Visual C++ 2015 produced before and after this change.
They are identical.