Annotate that we test with GHC version >= 7.4.1
We are relying on some implementation details of GHC, like the minimum
size of Int and language extensions. gbc/bond might work with other
Haskell compilers, but we don't test with them.

We don't list all the GHC versions that we end up using, as Travis CI
can change the specific versions on us.