Use scientific >= 0.3.4.6 for double serialization
Upgrade to scientific-0.3.4.6 or later so that when we serialize whole
numbers in JSON we don't get trailing ".0" values.

Before, 2 would serialize as "2", but 0 would serialize "0.0". Now, 2
serializes as "2", and 0 serializes as "0".

Since the golden tests have the results of JSON serialization commited,
we have to be on one side or the other of scientific-0.3.4.6. We're
after it now. (Unless, we needlessly complicate the tests with some sort
of custom comparisons or smudge/clean process.)

See also https://github.com/bos/aeson/issues/369 and
https://github.com/basvandijk/scientific/commit/0f28347b4a3221a741f76a48f18ffb9de961e856