Don't download and embed images for C# docs
With the `--self-contained` flag to Pandoc, we were downloading the
NuGet version badges at build time and embedding them in the generated
HTML as data URIs. We actually want those to be fetched anew each time
someone renders the page so they see the latest version, not the version
that was current when the page was generated.

This should also get rid of the build failures we were seeing when
shields.io was down.