Require explicit opt-in to use pre-compiled gbc
Before this change, the build would look for pre-compiled gbc in a lot
of places on the system. It would often find gbc from a prior install of
Bond. However, this frequently causes the build to fail. Now, it only
looks for pre-compiled gbc in the directory specified by the environment
variable BOND_GBC_PATH. Alternatively, if the CMake variable
BOND_GBC_PATH is set, that will be used directly.