java core: Fix float fields with int defaults.
This is valid bond and apparently valid in all target languages. We
were falling through a match that believed integer literals should be
suffixed with nothing or L based on the type of the field, and did not
expect to be called for float/double fields.

Java has no problem initializing
  * a long with an int literal
  * a float/double with an int/long
but does require that integer literals end in L if they do not fit in
the range of int, unlike C#. This change makes us decide whether to
append the L based on the value of the default and ignores the type of
the field.