java compat: Initial commit.
This decides whether to run Java compat and gets it set up properly on
Win/Lin/Mac. It does this based on the existence of bond.jar and
compat.jar, meaning that, on a fresh tree, you need to run cmake again
after building gbc and the two jars.

Currently, all implemented protocols fail with EOFExceptions, and all
unimplemented protocols report themselves as failures.