[c#] Switch to bond.snk for all strong-name signing (#505)
Previously, when when an official release of Bond was produced, a
Microsoft-internal key instead of bond.snk was used to strong-name sign
the assemblies. This means that locally produced assemblies and
Microsoft-produced assemblies have different strong names, making it
difficult to consume non-official releases in applications that require
strong-name signed assemblies.

This change switches to using bond.snk for all strong-name signing.

Official Microsoft release will still be Authenticode signed for
integrity. (Official release of Bond are already Authenticode signed.)

This is a breaking change, as it will change the strong name of the
assemblies.

Resolves https://github.com/Microsoft/bond/issues/414