[c++] Fix potential simultaneous codegen
In some CMakeLists.txt files, codegen was being run on the same .bond
files multiple times. Depending on build scheduling, this could result
in build failures (if the generation were to happen at the same time) or
spurious rebuilds (due to spurious rewrites of the generated files).

Now, .bond files that are shared across targets within the same
directory are generated once and compiled into a static library that is
then linked into all the consumers.

Since we're compiling the .bond files into a static library, we can take
advantage of the _apply.cpp precompilation as well.