[c++ grpc] Fix incorrect casts to void*
The bond::ext::gRPC::io_manager expects pointers to bond::ext::gRPC::io_manager_tag base when it receives void* tags. However, we are are directly passing a pointer to derived type instead of explicitly casting to base (which does not cause issues due to current layout of those objects).