[travis] Include gbc dependencies in CI image
With this commit, the package dependencies that gbc and gbc-tests use
will be built once with Stack when creating the CI image. Subsequent CI
builds will then just build the gbc and gbc-tests parts and link with
the already built dependencies.

Should a commit modify the dependencies of gbc/gbc-tests, Stack is smart
enough to figure this out, and it will build the missing dependencies,
until a newer CI image is used that contains the updated dependencies.

Stack and cabal don't allow us to just redirect their intermediate
output directories, so the build_init_srack.zsh and build.zsh scripts
now collaborate to symlink .stack-work so that it's shared between image
building and image execution.

Minor changes:

* Added BUILD_SOURCE_COMMIT Docker build argument. This needs to be set
  to a commit in the https://github.com/Microsoft/bond repository that
  is used to determine the dependencies to build.
* The STACK_ROOT environment variable is defined in the Dockerfile so
  that all invocations of stack--both during image build and during
  image execution--will share the same location.
* Moved the BOND_ROOT environment variable to Dockerfile so it's shared
  across all scripts and across image build and image execution.