[travis java] Switch to openjdk-8-jdk-headless
Use the "headless" JDK in the Travis CI image to avoid installing a
bunch of X11 (and related) packages that we don't need. This also
reduces the size of the image.

After getting this commit into master, the next step is to wait for a
new image to be built and then consume that image in .travis.yml.

[skip ci]