[c#] Throw instead of Debug.Assert in examples
Switch the C# examples to throw when the conditions they are
demonstrating are violated instead of using Debug.Assert.

Debug.Assert often pops up a UI dialog with the details. The example are
all console applications, and this UI dialog caused them to stall
instead of exiting with a non-zero exit code.