[appveyor] Test Boost 1.62+ only
This is part 1 of 2 in the process of dropping support for versions of
Boost older than two years. Boost 1.61+ or newer is now required. (Bond
will likely continue to work with older versions of Boost, but we'll no
longer actively test against them.)

We'd like to test with the current lowest version of Boost that Bond
supports, Boost 1.61. However, there is no Boost 1.61 available on
AppVeyor, so we start with the lowest version that is available in
AppVeyor and supported by Bond: 1.62.

This AppVeyor change needs to be made before the CMake minimum version
bump that we're planning to make with the Travis CI and README changes.

This is the AppVeyor-side change for
https://github.com/Microsoft/bond/issues/771