[c++] Overhaul bond::maybe
bond::maybe now has better support for rvalue-references and allocator
propagation.

This implementation also avoids default initializing an instance of T
when creating an maybe that holds nothing

bond::maybe is now implemented in terms of boost::optional. We should be
able to std::optional with minimal effort when Bond requires a minimum
of C++17.

Additional changes:

* Removed a dangerous implicit conversion operator from `bond::maybe<T>`
  to `const T&`. To access a `bond::maybe<T>` value, use
* Added noexcept variants of `bond::maybe<T>::value`.
* Added `bond::maybe<T>::emplace` to construct a maybe's value in place.
* Added various rvalue-reference and allocator-aware constructors and
  assignment operators.
* Added `bond::maybe<T>::operator==(const T&)` to comapre directly to
  instances of `T`.
* Improved test coverage of `set_value()`: it was only effectively
  constructing a new T the first time.
* Added tests for moved-from bond::maybe instances.
* Added tests for allocator propagation.