[c++] Rename scoped_allocator example
Rename the scoped allocator example from "scoped_allocator" to
"scoped_alloc" so that it doesn't conflict with the "scoped_allocator"
header from the C++ standard library.

When --scoped-alloc is passed to gbc, the generated _types.h header
has a `#include <scoped_allocator>`.

When using CMake to generate Makefiles, the directory where the _types.h
generated file is placed is the same directory where the final
executable is placed.

When rebuilding the scoped_allocator example, the preprocessor tries to
#include the previous generated executable.

Implementations of the preprocessor differ in how they handle "file" vs.
<file>. Some will search in the current directory for <file>, while some
will not.

Renaming avoids this problem instead of relying on the behavior of
<file>.