[travis] Handle ACR manifest fractional seconds
ACR manifests have started using fractional seconds in their timestamps.
This has broken the fixed format string parsing.

The [%f format specifier][1] will not work in this case, as it accepts
up to 6 digits, but ACR is sometimes sending 7 or more digits. It is
difficult to install Python modules in the environment where the clean
up script runs, so we're limited to vanilla Python 3.

Therefore, this change adds a method that strips out the fractional
seconds. We don't need fractional second precision for deciding whether
or not to prune an old image.

[1]: https://docs.python.org/3.5/library/datetime.html#strftime-and-strptime-behavior