Make _grpc.h types C++20 compatible
As of C++20, it is malformed to have an unnamed class used in typedef with a
base class. In the generated reflection code for a service, we were using the
following construct to create a reflectable entry for each field in a
struct.

```
struct service {
    typedef struct /* no name */ : ::bond::reflection::MethodTemplate<...> {} method_name;
}
```

Now, instead of using an unnamed class, we generate a named class and refer
to that:

```
struct service {
    typedef struct method_name_type : ::bond::reflection::MethodTemplate<...> {} field_name;
}
```

This commit is very similar in spirit to f1cb707134. See that commit for
even more details about what's going on.

Also adds a test that compiles a service that has method names that
conflict with the names in `MethodTemplate`. The gRPC++ unit test CMake
now has a single library with the compiled generated code.

Fixes https://github.com/microsoft/bond/issues/1027