[c#] Fix duplicate Compile item error with --grpc
When a project has multiple .bond files in `@BondCodegen` and specified
`--grpc` in `$BondOptions`, the C# code generation targets erroneously
added the generated files to the `@Compile` item multiple times.

This was caused because the `@_BondGen_Grpc_BondOptions` item
incorrectly used `%(BondCodegen.Identity)` when populating
`%DependentUpon`. Since `@_BondGen_Grpc_BondOptions` and `@BondCodegen`
are different items, this multiplied them together. A mapping was
intended instead, so the fix is to use
`%_BondGen_Grpc_BondOptions.Identity`. (This was likely a copy/paste bug
when `--grpc` support was added to the targets.)

Fixes https://github.com/microsoft/bond/issues/1050