[c#] Don't access Stream.Position in Release cfg.
Previously, `Stream.Position` was only accessed in the argument list of
`PushLengthCheck`. `PushLengthCheck` is a conditional method that only
exists in the DEBUG configuration. In the published RELEASE version of
Bond, the call to `PushLengthCheck` _and the evaluation of its
arguments_ was not present.

In b0fd4a15 ([c#] Fix handling of large container lengths, 2020-07-06)
the computation of the position was moved into a function local. This
caused `Stream.Position` to be accessed any time Compact Binary v2 was
used, causing serialization failures when using unseekable streams, like
those that ASP.NET Core uses for HTTP responses.

Move the access of `Stream.Position` back inside of the
`PushLengthCheck` argument list so that it is only accessed in the DEBUG
configuration to fix this regression.