Replace BOND_NORETURN with [[noreturn]]
Bond requires a C++11 compiler, The `[[noreturn]]` attribute is part of
C++11 and is supported by all of the minimum versions of the compilers
Bond supports.

Replace the `BOND_NORETURN` macro with the `[[noreturn]]` attribute.

The platform/compiler adapter macros are considered implementation
details of Bond. The removal of `BOND_NORETURN` is not considered a
breaking change.