Update Ubuntu image: use newer Git
The GitHub Actions workflows need a newer version of Git installed than
the version the Ubuntu 16.04 image comes with. In commit e1d1f39e ([ci]
Enable GitHub Actions for Linux CI, 2021-01-13), the Dockerfile that
creates the image we use was updated to install a newer version.

Switch the CI pipelines to this image. Remove GitHub Actions
step that installs the newer Git.