[docs] Switch to robocopy to mirror doc build output
We're having trouble using the Remove-Item/Copy-Item pair to synchronize
the documentation build output with the gh-pages branch.

Switch to robocopy, which has a /MIR (mirror) mode as well as
/XD (exclude directory).

Test $LASTEXITCODE instead of $?. Due to how PowerShell handles the
error stream, $? is set to false if the process writes to stderr, even
if the process exits with 0.