[ci] Skip GitHub workflow if changes don't affect the build
The GitHub workflow runs even if the changes don't affect the Linux
build.

Skip workflow runs for commits that don't affect the Linux build.

Negative path patterns are used in instead of paths-ignore: if anything
in paths-ignore matches, the workflow will be skipped, which is too
aggressive. Instead, the workflow should be skipped if the change solely
affects ignored files.